Fix special2Play issue
Fix the choosing possibilities for the main player (only possible to drop
a two when a special two is on play) by changing the attribute suit back
to String (it was changed incorrectly to Suits). Improved the getCardHandStr
function by separating if/else. Change handCards attribute to cardHand.
Print correctly valid cards on console. Change getSuitStr to getSuitOnPlay
to avoid confussion with getting a String. Take last card from the draw
pile instead of first.